Prevent rmrf if source is output. Proper exit code

+++ index.js
@@ -30,7 +30,7 @@ app
 let count = 0
 
 // Clear any existing output
-rimraf.sync(path.join(process.cwd(), app.output))
+if (app.source !== app.output) rimraf.sync(path.join(process.cwd(), app.output))
 
 // Read existing YAML file
 yaml.read(path.join(process.cwd(), app.source)).then(data => {
@@ -49,4 +49,5 @@ yaml.read(path.join(process.cwd(), app.source)).then(data => {
   console.log(colors.green(`Added ${count} environment variables to ${app.output}`))
 }).catch(error => {
   console.error(colors.red(error.message))
+  process.exit(1)
 })
